requirements.txt
Django==2.0.6
psycopg2
pytest-django
